email_agent/requirements.txt
azure-identity
msgraph-core
dotenv 
msgraph-sdk
